working on test for pandas io
